compiler/module: Catch panics properly.
Currently, if a compiler panics, the error is recorded into a local
variable defined in the scope of a Compile* function.  This variable
however has no direct relation to the return value of the function.
So if a panic actually happens, the last `return` statement is not
executed and the function just returns empty default values.

This behaviour results in errors being silently discurded, the
concequences of such errors are typically discovered much later during
execution (e.g., as index out of bounds panics in vm.go).

After this change, panics in Compile* functions are always propagated
to the caller as errors.